notebooklm-bridge/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
notebooklm-py==0.3.3
python-dotenv==1.0.0
httpx==0.27.0
